+++ packages/react-reconciler/src/ReactFiberRoot.js
@@ -16,10 +16,7 @@ import type {Interaction} from 'scheduler/src/Tracing';
 import {noTimeout} from './ReactFiberHostConfig';
 import {createHostRootFiber} from './ReactFiber';
 import {NoWork} from './ReactFiberExpirationTime';
-import {
-  enableSchedulerTracing,
-  enableNewScheduler,
-} from 'shared/ReactFeatureFlags';
+import {enableSchedulerTracing} from 'shared/ReactFeatureFlags';
 import {unstable_getThreadID} from 'scheduler/tracing';
 
 // TODO: This should be lifted into the renderer.
@@ -40,30 +37,10 @@ type BaseFiberRootProperties = {|
   // The currently active root fiber. This is the mutable root of the tree.
   current: Fiber,
 
-  // The following priority levels are used to distinguish between 1)
-  // uncommitted work, 2) uncommitted work that is suspended, and 3) uncommitted
-  // work that may be unsuspended. We choose not to track each individual
-  // pending level, trading granularity for performance.
-  //
-  // The earliest and latest priority levels that are suspended from committing.
-  earliestSuspendedTime: ExpirationTime,
-  latestSuspendedTime: ExpirationTime,
-  // The earliest and latest priority levels that are not known to be suspended.
-  earliestPendingTime: ExpirationTime,
-  latestPendingTime: ExpirationTime,
-  // The latest priority level that was pinged by a resolved promise and can
-  // be retried.
-  latestPingedTime: ExpirationTime,
-
   pingCache:
     | WeakMap<Thenable, Set<ExpirationTime>>
     | Map<Thenable, Set<ExpirationTime>>
     | null,
-
-  // If an error is thrown, and there are no more updates in the queue, we try
-  // rendering from the root one more time, synchronously, before handling
-  // the error.
-  didError: boolean,
 
   pendingCommitExpirationTime: ExpirationTime,
   // A finished work-in-progress HostRoot that's ready to be committed.
@@ -76,22 +53,19 @@ type BaseFiberRootProperties = {|
   pendingContext: Object | null,
   // Determines if we should attempt to hydrate on the initial mount
   +hydrate: boolean,
-  // Remaining expiration time on this root.
-  // TODO: Lift this into the renderer
-  nextExpirationTimeToWorkOn: ExpirationTime,
-  expirationTime: ExpirationTime,
   // List of top-level batches. This list indicates whether a commit should be
   // deferred. Also contains completion callbacks.
   // TODO: Lift this into the renderer
   firstBatch: Batch | null,
-  // Linked-list of roots
-  nextScheduledRoot: FiberRoot | null,
-
-  // New Scheduler fields
+  // Node returned by Scheduler.scheduleCallback
   callbackNode: *,
+  // Expiration of the callback associated with this root
   callbackExpirationTime: ExpirationTime,
+  // The earliest pending expiration time that exists in the tree
   firstPendingTime: ExpirationTime,
+  // The latest pending expiration time that exists in the tree
   lastPendingTime: ExpirationTime,
+  // The time at which a suspended component pinged the root to render again
   pingTime: ExpirationTime,
 |};
 
@@ -127,24 +101,11 @@ function FiberRootNode(containerInfo, hydrate) {
   this.pendingContext = null;
   this.hydrate = hydrate;
   this.firstBatch = null;
-
-  if (enableNewScheduler) {
-    this.callbackNode = null;
-    this.callbackExpirationTime = NoWork;
-    this.firstPendingTime = NoWork;
-    this.lastPendingTime = NoWork;
-    this.pingTime = NoWork;
-  } else {
-    this.earliestPendingTime = NoWork;
-    this.latestPendingTime = NoWork;
-    this.earliestSuspendedTime = NoWork;
-    this.latestSuspendedTime = NoWork;
-    this.latestPingedTime = NoWork;
-    this.didError = false;
-    this.nextExpirationTimeToWorkOn = NoWork;
-    this.expirationTime = NoWork;
-    this.nextScheduledRoot = null;
-  }
+  this.callbackNode = null;
+  this.callbackExpirationTime = NoWork;
+  this.firstPendingTime = NoWork;
+  this.lastPendingTime = NoWork;
+  this.pingTime = NoWork;
 
   if (enableSchedulerTracing) {
     this.interactionThreadID = unstable_getThreadID();
