+++ packages/react-reconciler/src/SchedulerWithReactIntegration.js
@@ -10,8 +10,11 @@
 // Intentionally not named imports because Rollup would use dynamic dispatch for
 // CommonJS interop named imports.
 import * as Scheduler from 'scheduler';
-
-import {disableYielding} from 'shared/ReactFeatureFlags';
+import {__interactionsRef} from 'scheduler/tracing';
+import {
+  disableYielding,
+  enableSchedulerTracing,
+} from 'shared/ReactFeatureFlags';
 import invariant from 'shared/invariant';
 
 const {
@@ -27,6 +30,20 @@ const {
   unstable_LowPriority: Scheduler_LowPriority,
   unstable_IdlePriority: Scheduler_IdlePriority,
 } = Scheduler;
+
+if (enableSchedulerTracing) {
+  // Provide explicit error message when production+profiling bundle of e.g.
+  // react-dom is used with production (non-profiling) bundle of
+  // scheduler/tracing
+  invariant(
+    __interactionsRef != null && __interactionsRef.current != null,
+    'It is not supported to run the profiling version of a renderer (for ' +
+      'example, `react-dom/profiling`) without also replacing the ' +
+      '`scheduler/tracing` module with `scheduler/tracing-profiling`. Your ' +
+      'bundler might have a setting for aliasing both modules. Learn more at ' +
+      'http://fb.me/react-profiling',
+  );
+}
 
 export opaque type ReactPriorityLevel = 99 | 98 | 97 | 96 | 95 | 90;
 export type SchedulerCallback = (isSync: boolean) => SchedulerCallback | null;
