+++ packages/react-reconciler/src/__tests__/ReactLazy-test.internal.js
@@ -5,7 +5,6 @@ let Scheduler;
 let ReactFeatureFlags;
 let Suspense;
 let lazy;
-let enableNewScheduler;
 
 describe('ReactLazy', () => {
   beforeEach(() => {
@@ -19,7 +18,6 @@ describe('ReactLazy', () => {
     lazy = React.lazy;
     ReactTestRenderer = require('react-test-renderer');
     Scheduler = require('scheduler');
-    enableNewScheduler = ReactFeatureFlags.enableNewScheduler;
   });
 
   function Text(props) {
@@ -487,13 +485,7 @@ describe('ReactLazy', () => {
 
     await Promise.resolve();
 
-    if (enableNewScheduler) {
-      // The new scheduler pings in a separate task
-      expect(Scheduler).toHaveYielded([]);
-    } else {
-      // The old scheduler pings synchronously
-      expect(Scheduler).toHaveYielded(['UNSAFE_componentWillMount: A', 'A1']);
-    }
+    expect(Scheduler).toHaveYielded([]);
 
     root.update(
       <Suspense fallback={<Text text="Loading..." />}>
@@ -501,19 +493,7 @@ describe('ReactLazy', () => {
       </Suspense>,
     );
 
-    if (enableNewScheduler) {
-      // Because this ping happens in a new task, the ping and the update
-      // are batched together
-      expect(Scheduler).toHaveYielded(['UNSAFE_componentWillMount: A', 'A2']);
-    } else {
-      // The old scheduler must do two separate renders, no batching.
-      expect(Scheduler).toHaveYielded([
-        'UNSAFE_componentWillReceiveProps: A -> A',
-        'UNSAFE_componentWillUpdate: A -> A',
-        'A2',
-      ]);
-    }
-
+    expect(Scheduler).toHaveYielded(['UNSAFE_componentWillMount: A', 'A2']);
     expect(root).toMatchRenderedOutput('A2');
 
     root.update(
