+++ packages/react-reconciler/src/__tests__/ReactSchedulerIntegration-test.internal.js
@@ -26,7 +26,6 @@ describe('ReactSchedulerIntegration', () => {
     jest.resetModules();
     ReactFeatureFlags = require('shared/ReactFeatureFlags');
     ReactFeatureFlags.debugRenderPhaseSideEffectsForStrictMode = false;
-    ReactFeatureFlags.enableNewScheduler = true;
     React = require('react');
     ReactNoop = require('react-noop-renderer');
     Scheduler = require('scheduler');
