+++ packages/react-reconciler/src/__tests__/ReactSuspense-test.internal.js
@@ -5,7 +5,6 @@ let Scheduler;
 let ReactCache;
 let Suspense;
 let act;
-let enableNewScheduler;
 
 let TextResource;
 let textResourceShouldFail;
@@ -23,7 +22,6 @@ describe('ReactSuspense', () => {
     act = ReactTestRenderer.act;
     Scheduler = require('scheduler');
     ReactCache = require('react-cache');
-    enableNewScheduler = ReactFeatureFlags.enableNewScheduler;
 
     Suspense = React.Suspense;
 
@@ -267,11 +265,7 @@ describe('ReactSuspense', () => {
 
     await LazyClass;
 
-    if (enableNewScheduler) {
-      expect(Scheduler).toFlushExpired(['Hi', 'Did mount: Hi']);
-    } else {
-      expect(Scheduler).toHaveYielded(['Hi', 'Did mount: Hi']);
-    }
+    expect(Scheduler).toFlushExpired(['Hi', 'Did mount: Hi']);
     expect(root).toMatchRenderedOutput('Hi');
   });
 
@@ -400,24 +394,13 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(100);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B:1]']);
-        expect(Scheduler).toFlushExpired([
-          'B:1',
-          'Unmount [Loading...]',
-          // Should be a mount, not an update
-          'Mount [B:1]',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [B:1]',
-          'B:1',
-          'Unmount [Loading...]',
-          // Should be a mount, not an update
-          'Mount [B:1]',
-        ]);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [B:1]']);
+      expect(Scheduler).toFlushExpired([
+        'B:1',
+        'Unmount [Loading...]',
+        // Should be a mount, not an update
+        'Mount [B:1]',
+      ]);
       expect(root).toMatchRenderedOutput('AB:1C');
 
       instance.setState({step: 2});
@@ -430,21 +413,12 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(100);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B:2]']);
-        expect(Scheduler).toFlushExpired([
-          'B:2',
-          'Unmount [Loading...]',
-          'Update [B:2]',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [B:2]',
-          'B:2',
-          'Unmount [Loading...]',
-          'Update [B:2]',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [B:2]']);
+      expect(Scheduler).toFlushExpired([
+        'B:2',
+        'Unmount [Loading...]',
+        'Update [B:2]',
+      ]);
       expect(root).toMatchRenderedOutput('AB:2C');
     });
 
@@ -477,13 +451,8 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
-        expect(Scheduler).toFlushExpired(['A']);
-      } else {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]', 'A']);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
+      expect(Scheduler).toFlushExpired(['A']);
       expect(root).toMatchRenderedOutput('Stateful: 1A');
 
       root.update(<App text="B" />);
@@ -500,13 +469,8 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
-        expect(Scheduler).toFlushExpired(['B']);
-      } else {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B]', 'B']);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
+      expect(Scheduler).toFlushExpired(['B']);
       expect(root).toMatchRenderedOutput('Stateful: 2B');
     });
 
@@ -547,12 +511,8 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
-        expect(Scheduler).toFlushExpired(['A']);
-      } else {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]', 'A']);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
+      expect(Scheduler).toFlushExpired(['A']);
       expect(root).toMatchRenderedOutput('Stateful: 1A');
 
       root.update(<App text="B" />);
@@ -576,13 +536,8 @@ describe('ReactSuspense', () => {
 
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
-        expect(Scheduler).toFlushExpired(['B']);
-      } else {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B]', 'B']);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
+      expect(Scheduler).toFlushExpired(['B']);
       expect(root).toMatchRenderedOutput('Stateful: 2B');
     });
 
@@ -664,16 +619,8 @@ describe('ReactSuspense', () => {
       expect(Scheduler).toHaveYielded(['Suspend! [A]', 'Loading...']);
       jest.advanceTimersByTime(500);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
-        expect(Scheduler).toFlushExpired(['A', 'Did commit: A']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [A]',
-          'A',
-          'Did commit: A',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
+      expect(Scheduler).toFlushExpired(['A', 'Did commit: A']);
     });
 
     it('retries when an update is scheduled on a timed out tree', () => {
@@ -756,43 +703,25 @@ describe('ReactSuspense', () => {
         'Loading...',
       ]);
       expect(Scheduler).toFlushAndYield([]);
+
       jest.advanceTimersByTime(1000);
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Child 1]']);
-        expect(Scheduler).toFlushExpired([
-          'Child 1',
-          'Suspend! [Child 2]',
-          'Suspend! [Child 3]',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Child 1]',
-          'Child 1',
-          'Suspend! [Child 2]',
-          'Suspend! [Child 3]',
-        ]);
-      }
+
+      expect(Scheduler).toHaveYielded(['Promise resolved [Child 1]']);
+      expect(Scheduler).toFlushExpired([
+        'Child 1',
+        'Suspend! [Child 2]',
+        'Suspend! [Child 3]',
+      ]);
+
       jest.advanceTimersByTime(1000);
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Child 2]']);
-        expect(Scheduler).toFlushExpired(['Child 2', 'Suspend! [Child 3]']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Child 2]',
-          'Child 2',
-          'Suspend! [Child 3]',
-        ]);
-      }
+
+      expect(Scheduler).toHaveYielded(['Promise resolved [Child 2]']);
+      expect(Scheduler).toFlushExpired(['Child 2', 'Suspend! [Child 3]']);
+
       jest.advanceTimersByTime(1000);
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Child 3]']);
-        expect(Scheduler).toFlushExpired(['Child 3']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Child 3]',
-          'Child 3',
-        ]);
-      }
+
+      expect(Scheduler).toHaveYielded(['Promise resolved [Child 3]']);
+      expect(Scheduler).toFlushExpired(['Child 3']);
       expect(root).toMatchRenderedOutput(
         ['Child 1', 'Child 2', 'Child 3'].join(''),
       );
@@ -852,15 +781,8 @@ describe('ReactSuspense', () => {
       expect(root).toMatchRenderedOutput('Loading...');
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 0]']);
-        expect(Scheduler).toFlushExpired(['Tab: 0']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Tab: 0]',
-          'Tab: 0',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 0]']);
+      expect(Scheduler).toFlushExpired(['Tab: 0']);
       expect(root).toMatchRenderedOutput('Tab: 0 + sibling');
 
       act(() => setTab(1));
@@ -872,16 +794,8 @@ describe('ReactSuspense', () => {
       expect(root).toMatchRenderedOutput('Loading...');
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 1]']);
-        expect(Scheduler).toFlushExpired(['Tab: 1']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Tab: 1]',
-          'Tab: 1',
-        ]);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 1]']);
+      expect(Scheduler).toFlushExpired(['Tab: 1']);
       expect(root).toMatchRenderedOutput('Tab: 1 + sibling');
 
       act(() => setTab(2));
@@ -893,16 +807,8 @@ describe('ReactSuspense', () => {
       expect(root).toMatchRenderedOutput('Loading...');
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 2]']);
-        expect(Scheduler).toFlushExpired(['Tab: 2']);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [Tab: 2]',
-          'Tab: 2',
-        ]);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [Tab: 2]']);
+      expect(Scheduler).toFlushExpired(['Tab: 2']);
       expect(root).toMatchRenderedOutput('Tab: 2 + sibling');
     });
 
@@ -939,13 +845,8 @@ describe('ReactSuspense', () => {
       expect(Scheduler).toHaveYielded(['Suspend! [A:0]', 'Loading...']);
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A:0]']);
-        expect(Scheduler).toFlushExpired(['A:0']);
-      } else {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A:0]', 'A:0']);
-      }
-
+      expect(Scheduler).toHaveYielded(['Promise resolved [A:0]']);
+      expect(Scheduler).toFlushExpired(['A:0']);
       expect(root).toMatchRenderedOutput('A:0');
 
       act(() => setStep(1));
@@ -982,65 +883,35 @@ describe('ReactSuspense', () => {
       // Resolve A
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
-        expect(Scheduler).toFlushExpired([
-          'A',
-          // The promises for B and C have now been thrown twice
-          'Suspend! [B]',
-          'Suspend! [C]',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [A]',
-          'A',
-          // The promises for B and C have now been thrown twice
-          'Suspend! [B]',
-          'Suspend! [C]',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [A]']);
+      expect(Scheduler).toFlushExpired([
+        'A',
+        // The promises for B and C have now been thrown twice
+        'Suspend! [B]',
+        'Suspend! [C]',
+      ]);
 
       // Resolve B
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
-        expect(Scheduler).toFlushExpired([
-          // Even though the promise for B was thrown twice, we should only
-          // re-render once.
-          'B',
-          // The promise for C has now been thrown three times
-          'Suspend! [C]',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [B]',
-          // Even though the promise for B was thrown twice, we should only
-          // re-render once.
-          'B',
-          // The promise for C has now been thrown three times
-          'Suspend! [C]',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [B]']);
+      expect(Scheduler).toFlushExpired([
+        // Even though the promise for B was thrown twice, we should only
+        // re-render once.
+        'B',
+        // The promise for C has now been thrown three times
+        'Suspend! [C]',
+      ]);
 
       // Resolve C
       jest.advanceTimersByTime(1000);
 
-      if (enableNewScheduler) {
-        expect(Scheduler).toHaveYielded(['Promise resolved [C]']);
-        expect(Scheduler).toFlushExpired([
-          // Even though the promise for C was thrown three times, we should only
-          // re-render once.
-          'C',
-        ]);
-      } else {
-        expect(Scheduler).toHaveYielded([
-          'Promise resolved [C]',
-          // Even though the promise for C was thrown three times, we should only
-          // re-render once.
-          'C',
-        ]);
-      }
+      expect(Scheduler).toHaveYielded(['Promise resolved [C]']);
+      expect(Scheduler).toFlushExpired([
+        // Even though the promise for C was thrown three times, we should only
+        // re-render once.
+        'C',
+      ]);
     });
 
     it('#14162', () => {
