+++ packages/shared/ReactFeatureFlags.js
@@ -65,9 +65,5 @@ export const warnAboutDeprecatedSetNativeProps = false;
 // Experimental React Events support. Only used in www builds for now.
 export const enableEventAPI = false;
 
-// Enables rewritten version of ReactFiberScheduler. Added in case we need to
-// quickly revert it.
-export const enableNewScheduler = false;
-
 // New API for JSX transforms to target - https://github.com/reactjs/rfcs/pull/107
 export const enableJSXTransformAPI = false;
