+++ packages/shared/forks/ReactFeatureFlags.test-renderer.www.js
@@ -26,7 +26,6 @@ export const warnAboutDeprecatedSetNativeProps = false;
 export const disableJavaScriptURLs = false;
 export const disableYielding = false;
 export const enableEventAPI = true;
-export const enableNewScheduler = false;
 export const enableJSXTransformAPI = true;
 
 // Only used in www builds.
