--- scripts/jest/config.source-new-scheduler.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const baseConfig = require('./config.base');
-
-module.exports = Object.assign({}, baseConfig, {
-  setupFiles: [
-    ...baseConfig.setupFiles,
-    require.resolve('./setupNewScheduler.js'),
-    require.resolve('./setupHostConfigs.js'),
-  ],
-});
