--- scripts/jest/setupNewScheduler.js
@@ -1,7 +1,0 @@
-'use strict';
-
-jest.mock('shared/ReactFeatureFlags', () => {
-  const ReactFeatureFlags = require.requireActual('shared/ReactFeatureFlags');
-  ReactFeatureFlags.enableNewScheduler = true;
-  return ReactFeatureFlags;
-});
